bug(board): use header pagination links for contents
This commit makes changes in both packages and the legacy SDK to make
the pagination tests for contents more reliable.  Now that the server is
providing pagination in the header as 'link', it is easier to manage the
size of the chunks for gets and adds.

This commit also changes the size of the chunks for adds to 150 contents
per chunk.  If a client wants to add 300 contents in one transaction,
then the it would split it into 2 chunks of 150.