refactor(@ciscospark/spark-core): make imports transform-friendly
logger.js imported a few things in a way that made it hard to write a
codemod that would behave correctly. This tweaks the import statement(s)
just enough to make writing a codemod way easier. By committing this
before applying the mod instead of after, it'll make it easier to have
the programatic changes reviewed and removed, then rebase against master
and reapply the mod.