fix(plugin-people): modify people.list to finish batching when notFoundIds exists
The total batched PII requests are equal to the number of returned PIIs (body.items) and the list of missing PIIs (body.notFoundIds). people.list must resolve both items and notFoundIds to complete the batch and ultimately resolve the requested PIIs.
Previously notFoundIds was ignored and the batch always failed with a timeout.