fix(plugin-messages): update message created event  handling
Update how message:created events are shaped so that they properly
populate the text, markdown, and files fields. Clean up testing
so that it handles this properly, and adjust testing statements
to utilize the proper 'files' field instead of 'file'. Update
documentation for messages.listen() to identify how to retrieve
the html field data to satisfy the related internal case.