docs(plugin-memberships): support updating isRoomHidden attribute
Applications may now call memberships.update() and set the isRoomHidden attribute of the user's
membership to true or false for a 1-1 space that the user belongs to.
Setting it to false is effectively the same thing as a user choosing the "leave space" option in the
Webex Teams client for a 1-1 space.   The client will no longer display that 1-1 space in the user's
recents list, but if the other user sends a message it becomes "unhidden" and all the previous
conversation history is available.
When a space's hidden status changes for a user, a memberships:updated event is generated for
that user.