fix(plugin-meetings): fix various scenarios related to screen sharing
This commit fixes a number of issues, most of them related to scenarios
where a meeting participant starts sharing while someone else is already
doing a screen share.

I've also introduced a new public field shareStatus that should give the
app full information about the current share status. It is needed
because otherwise if the app relies only on the notifications, it will
not get the meeting:startedSharingRemote event if page is loaded after
someone already started sharing (the SDK sends startedSharingRemote
event before the app registers for it).

Now the isSharing property is deprecated and I've changed the code so
that internally we don't use isSharing anywhere, so it can be easily
removed in the future.

Fixes #SPARK-222423
Fixes #SPARK-222430
Fixes #SPARK-222426
Fixes #SPARK-220647