feat(meetings): custom size remote video (coming soon)
fixes: SPARK-227135

Extending the changeVideoLayout API so that the app
can send renderInfo to the backend. The actual sending
of the request is debounced, because /control endpoint
has a rate limit of 24 per second and it's very likely
that developers will call this API very frequently in
some scenarios (for example on window size change).

For now the renderInfo is ignored by the backend. Backend
teams are working on adding support for it.