fix: support for connecting to realtime channel for federated boards
Federated boards are ones that live in a cluster other than the active
user's default cluster. Registration of a websocket for these boards
must happen through a channel-aware registration API, rather than the
deprecated cluster-wide registration API.

This change uses that approach when the board.useFederatedRegistration
configuration is set to 'true', and uses the current approach when
that config is not set. So, clients wishing for the federation-capable
approach must set the config option in the config tree when constructing
their Webex object.

This could be further developed to use a feature toggle for rollout, but
any such improvements should still respect the configuration value as a
default, so that clients other than the web client can enable the
federated behavior.

This addresses #2155.