fix: downscopeing issues
Scopes defined by the app and the scopes from the idbroker can mismatch and
cause "invalid scope" error during downscoping.

It can happen when a new release of the app define more scopes,
but the user did not logout and in again to get the new scopes from idbroker.
Unfortunately super token refresh does not update the scopes.

solution:
Exclude invalid scopes from the downscope request (and log a warning),
except when the downscope called with only invalid scopes,
in this case we let handle the error by the downscope function
and fallback to the super token.

This should greatly reduce the the number of 400 error during downscope.

issue: SPARK-474816